+++ .settings/org.eclipse.php.core.prefs
@@ -1,0 +1,4 @@
+eclipse.preferences.version=1
+include_path=0;/laval.io
+phpVersion=php5.6
+use_asp_tags_as_php=false
