+++ .settings/org.eclipse.wst.common.project.facet.core.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<faceted-project>
+  <fixed facet="php.component"/>
+  <fixed facet="php.core.component"/>
+  <installed facet="php.core.component" version="1"/>
+  <installed facet="php.component" version="5.6"/>
+</faceted-project>
