+++ App/Model/User.php
@@ -1,0 +1,9 @@
+<?php
+
+/*
+ * Gestion des données utilisateur
+ */
+class User extends Models{
+	
+
+}
