+++ App/View/users/index.php
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="fr">
+<head>
+	<meta charset="UTF-8">
+	<title>LAVAL</title>
+</head>
+<body>
+	
+</body>
+</html>
