+++ App/controllers/BriqueController.php
@@ -1,0 +1,1 @@
+<?php
