+++ App/controllers/UsersController.php
@@ -1,0 +1,11 @@
+<?php
+
+/*
+*  Gestion de la logique utilisateur
+*/
+
+class UsersController extends Controller{
+
+
+
+}
