+++ App/core/Bootstrap.php
@@ -1,0 +1,15 @@
+<?php
+
+  class Bootstrap {
+  	
+  	public function __construct(){
+  		$url = $_GET['url'];
+  		$url = trim($url, '/');
+  		$url = explode($url);
+  	}
+     
+
+
+  }
+
+ ?>
