+++ App/core/Controller.php
@@ -1,0 +1,1 @@
+<?php
