+++ App/core/Model.php
@@ -1,0 +1,8 @@
+<?php
+  class Model {
+    
+    public function __construct() {
+
+    }
+  }
+ ?>
