+++ App/core/Router.php
@@ -1,0 +1,5 @@
+<?php
+
+ class Route {
+ 	
+ }
