--- Controller/UsersController.php
@@ -1,11 +1,0 @@
-<?php
-
-/*
-*  Gestion de la logique utilisateur
-*/
-
-class UsersController{
-
-
-
-}
