--- Model/User.php
@@ -1,9 +1,0 @@
-<?php
-
-/*
- * Gestion des données utilisateur
- */
-class User{
-
-
-}
