--- View/index.php
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-	<meta charset="UTF-8">
-	<title>LAVAL</title>
-</head>
-<body>
-	
-</body>
-</html>
