+++ index.php
@@ -1,0 +1,4 @@
+<?php 
+
+	require "vendor/autolaod.php";
+
